Refactor API URLs to use BASE_API_URL env variable
Replaced service-specific environment variables with BASE_API_URL across frontend API routes and microservices for consistency and easier configuration. Also renamed API route directories from plural to singular for client and invoice resources, and added logging to the fetcher utility.
