+++ frontend/app/api/auth/me/route.ts
@@ -2,7 +2,7 @@ import { forward } from "@/lib/fetcher"
 import { NextResponse } from "next/server"
 import { authFrom } from "../../_utils/auth"
 
-const AUTH_URL = process.env.AUTH_SERVICE_URL!
+const AUTH_URL = process.env.BASE_API_URL!
 
 export async function GET(req: Request) {
   try {
