+++ frontend/app/api/invoice/[id]/notify/route.ts
@@ -2,7 +2,7 @@ import { NextResponse } from "next/server"
 import { authFrom, errorJson } from "../../../_utils/auth"
 import { forward } from "@/lib/fetcher"
 
-const BASE = process.env.INVOICE_SERVICE_URL!
+const BASE = process.env.BASE_API_URL!
 
 export async function POST(_: Request, { params }: { params: { id: string } }) {
   try {
