+++ frontend/app/api/invoice/route.ts
@@ -1,19 +1,16 @@
 import { NextResponse } from "next/server"
-import { authFrom, json, errorJson } from "../_utils/auth"
+
+import { authFrom, errorJson } from "../_utils/auth"
 import { forward } from "@/lib/fetcher"
 
-const BASE = process.env.CLIENT_SERVICE_URL!
+const BASE = process.env.BASE_API_URL!
 
-// GET /api/client?search=... (optional passthrough)
 export async function GET(req: Request) {
-  const url = new URL(req.url)
-  const search = url.searchParams.get("search")
-  const target = search
-    ? `${BASE}/api/client?search=${encodeURIComponent(search)}`
-    : `${BASE}/api/client`
-
   try {
-    const data = await forward(target, { headers: authFrom(req) })
+    const auth = authFrom(req)
+    const data = await forward(`${BASE}/api/invoice`, {
+      headers: Object.keys(auth).length > 0 ? auth : {},
+    })
     return NextResponse.json(data)
   } catch (e: any) {
     const { status, body } = errorJson(e)
@@ -21,14 +18,16 @@ export async function GET(req: Request) {
   }
 }
 
-// POST /api/client
 export async function POST(req: Request) {
   try {
     const body = await req.json()
-    const data = await forward(`${BASE}/api/client`, {
+    const data = await forward(`${BASE}/api/invoice`, {
       method: "POST",
-      headers: authFrom(req),
-      ...json(body),
+      headers: {
+        ...authFrom(req),
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(body),
     })
     return NextResponse.json(data, { status: 201 })
   } catch (e: any) {
