+++ frontend/lib/fetcher.ts
@@ -43,6 +43,8 @@ export async function forward<T = any>(
 
   let res: Response
   try {
+    console.log(`[fetcher] ${init?.method || "GET"} ${url}`)
+
     res = await fetch(url, {
       cache: "no-store",
       ...init,
