+++ microservices/auth-service/.http
@@ -1,4 +1,4 @@
-@baseUrl = http://localhost:3002/api/auth
+@baseUrl = https://zxlnv1v3xb.execute-api.us-east-1.amazonaws.com/api/auth
 @TOKEN= eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY4OTUwZDU1OTAyY2MzZTdlYWJhNzQ5NyIsImVtYWlsIjoicmlzaGhpQGV4YW1wbGUxLmNvbSIsImlhdCI6MTc1NDU5OTI0Nn0.YQ8soDNBIhFuI-0U0b-7F3HIDau_O7UTRg_Vwe-my9k
 
 ### REGISTER a new user
